flask_requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-google-genai
langchain-community
youtube_transcript_api
chromadb
google-genai
PyPDF2
python-docx
scikit-learn
numpy
werkzeug
requests